chore: support requestStream history
