+++ packages/ai-native/src/browser/components/utils.ts
@@ -1,8 +1,9 @@
 import { ITextMessageProps } from 'react-chat-elements';
 
 import { CODICON_OWNER, getExternalIcon } from '@opensumi/ide-core-browser';
+import { ChatMessageRole } from '@opensumi/ide-core-common/lib/types/ai-native';
 
-import { ChatMessageRole, IChatReplyFollowup, ISampleQuestions } from '../../common/index';
+import { IChatReplyFollowup, ISampleQuestions } from '../../common/index';
 
 export interface MessageData extends Pick<ITextMessageProps, 'id' | 'position' | 'className' | 'title'> {
   role: ChatMessageRole;
