+++ packages/ai-native/src/browser/model/msg-history-manager.ts
@@ -1,20 +1,7 @@
 import { Injectable } from '@opensumi/di';
-import { Dispatcher, Disposable, Emitter, Event, uuid } from '@opensumi/ide-core-common';
-
-import { ChatMessageRole, IChatMessage } from '../../common/index';
-
-export interface IHistoryChatMessage extends IChatMessage {
-  id: string;
-  order: number;
-
-  type?: 'string' | 'component';
-  relationId?: string;
-  component?: React.ReactNode;
-  componentProps?: { [key in string]: any };
-
-  agentId?: string;
-  agentCommand?: string;
-}
+import { Disposable, Emitter, Event, uuid } from '@opensumi/ide-core-common';
+import { ChatMessageRole } from '@opensumi/ide-core-common/lib/types/ai-native';
+import { IHistoryChatMessage } from '@opensumi/ide-core-common/lib/types/ai-native';
 
 type IExcludeMessage = Omit<IHistoryChatMessage, 'id' | 'order'>;
 
@@ -55,14 +42,7 @@ export class MsgHistoryManager extends Disposable {
     return Array.from(this.messageMap.values()).sort((a, b) => a.order - b.order);
   }
 
-  public addUserMessage(message: Omit<IExcludeMessage, 'role'>): string {
-    return this.doAddMessage({
-      ...message,
-      role: ChatMessageRole.User,
-    });
-  }
-
-  public addAgentMessage(message: Required<Pick<IExcludeMessage, 'agentId' | 'agentCommand' | 'content'>>): string {
+  public addUserMessage(message: Required<Pick<IExcludeMessage, 'agentId' | 'agentCommand' | 'content'>>): string {
     return this.doAddMessage({
       ...message,
       role: ChatMessageRole.User,
