+++ packages/extension/src/common/sumi/chat-agents.ts
@@ -7,10 +7,10 @@ import type {
   IChatAgentResult,
   IChatAgentWelcomeMessage,
   IChatFollowup,
-  IChatMessage,
   IChatReplyFollowup,
 } from '@opensumi/ide-ai-native/lib/common';
 import type { IChatAsyncContent, IChatContent, IChatProgress } from '@opensumi/ide-core-common';
+import type { IChatMessage } from '@opensumi/ide-core-common/lib/types/ai-native';
 
 export interface IExtensionChatAgentMetadata extends IChatAgentMetadata {
   hasSlashCommands?: boolean;
